Exception refactor little fix
