+++ src/main/java/faang/school/projectservice/exception/NotUniqueProjectException.java
@@ -1,0 +1,8 @@
+package faang.school.projectservice.exception;
+
+public class NotUniqueProjectException extends RuntimeException {
+
+    public NotUniqueProjectException(String message) {
+        super(message);
+    }
+}
