+++ src/main/java/faang/school/projectservice/validator/ProjectValidator.java
@@ -1,6 +1,7 @@
 package faang.school.projectservice.validator;
 
-import faang.school.projectservice.exception.EntityNotFoundException;
+import faang.school.projectservice.dto.ProjectDto;
+import faang.school.projectservice.exception.NotUniqueProjectException;
 import faang.school.projectservice.repository.ProjectRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -12,11 +13,14 @@ import org.springframework.stereotype.Component;
 public class ProjectValidator {
     private final ProjectRepository projectRepository;
 
-    public void validateUniqueProject(String name, Long ownerId) {
+    public void validateUniqueProject(ProjectDto dto) {
+        Long ownerId = dto.getOwnerId();
+        String name = dto.getName();
+
         if (projectRepository.existsByOwnerUserIdAndName(ownerId, name)) {
             log.error("Project '{}' with ownerId #{} already exists.", name, ownerId);
-            throw new EntityNotFoundException("Project '" + name + "' with ownerId #" + ownerId + " already exists.");
+            throw new NotUniqueProjectException("Project '" + name + "' with ownerId #" + ownerId + " already exists.");
         }
-        log.info("Project '{}' with ownerId #{} does not exist. Can be created.", name, ownerId);
+        log.info("Project '{}' with ownerId #{} unique and can be created.", name, ownerId);
     }
 }
